refactor findPrincipalId method into common util class.
